src/requirements.txt
discord.py==2.3.2
public-ip==0.12
